[CodeQuality] Move AssertClassToThisAssertRector hook from Class_ to ClassMethod
The rule only ever rewrote Assert::assert*() calls inside class methods, so hooking on ClassMethod is more precise and avoids re-traversing the whole class.

Static methods are now skipped with an early return instead of a DONT_TRAVERSE_CURRENT_AND_CHILDREN check inside the traversal callback; static closures and arrow functions are still skipped during traversal.
