+++ rules-tests/CodeQuality/Rector/ClassMethod/AssertClassToThisAssertRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\CodeQuality\Rector\Class_\AssertClassToThisAssertRector;
+use Rector\PHPUnit\CodeQuality\Rector\ClassMethod\AssertClassToThisAssertRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(AssertClassToThisAssertRector::class);
